AccountClient removed - e2e tests are passing
