+++ starknet_py/net/__init__.py
@@ -1,1 +1,1 @@
-from .account.account_client import AccountClient, KeyPair
+from starknet_py.net.signer.stark_curve_signer import KeyPair
