+++ starknet_py/tests/e2e/account/account_test.py
@@ -5,11 +5,8 @@ from starkware.starknet.public.abi import get_selector_from_name
 
 from starknet_py.contract import Contract
 from starknet_py.hash.address import compute_address
-from starknet_py.net import AccountClient, KeyPair
-from starknet_py.net.account._account_proxy import AccountProxy
 from starknet_py.net.account.account import Account
 from starknet_py.net.account.base_account import BaseAccount
-from starknet_py.net.client import Client
 from starknet_py.net.client_errors import ClientError
 from starknet_py.net.client_models import (
     Call,
@@ -137,9 +134,9 @@ async def test_get_class_hash_at(map_contract, account):
 
 
 @pytest.mark.asyncio()
-async def test_get_nonce(gateway_account, base_account_deploy_map_contract):
+async def test_get_nonce(gateway_account, map_contract):
     nonce = await gateway_account.get_nonce()
-    address = base_account_deploy_map_contract.address
+    address = map_contract.address
 
     tx = await gateway_account.execute(
         Call(
@@ -197,7 +194,7 @@ async def test_sign_deploy_account_transaction(gateway_account):
 
 @pytest.mark.asyncio
 async def test_deploy_account(
-    client, deploy_account_details_factory, base_account_deploy_map_contract
+    client, deploy_account_details_factory, map_contract
 ):
     address, key_pair, salt, class_hash = await deploy_account_details_factory.get()
 
@@ -220,7 +217,7 @@ async def test_deploy_account(
     # Test making a tx
     res = await account.execute(
         calls=Call(
-            to_addr=base_account_deploy_map_contract.address,
+            to_addr=map_contract.address,
             selector=get_selector_from_name("put"),
             calldata=[30, 40],
         ),
@@ -374,37 +371,17 @@ async def test_sign_declare_tx_for_fee_estimation(account, map_compiled_contract
     await account.client.wait_for_tx(result.transaction_hash)
 
 
-def _account_by_type(
-    *, address: int, client: Client, key_pair: KeyPair, account_type: str
-) -> BaseAccount:
-    if account_type == "proxy":
-        return AccountProxy(
-            AccountClient(
-                address=address,
-                client=client,
-                key_pair=key_pair,
-                chain=StarknetChainId.TESTNET,
-                supported_tx_version=1,
-            )
-        )
+@pytest.mark.asyncio
+async def test_sign_deploy_account_tx_for_fee_estimation(
+    client, deploy_account_details_factory
+):
+    address, key_pair, salt, class_hash = await deploy_account_details_factory.get()
 
-    return Account(
+    account = Account(
         address=address,
         client=client,
         key_pair=key_pair,
         chain=StarknetChainId.TESTNET,
-    )
-
-
-@pytest.mark.parametrize("account_type", ("base", "proxy"))
-@pytest.mark.asyncio
-async def test_sign_deploy_account_tx_for_fee_estimation(
-    client, deploy_account_details_factory, account_type
-):
-    address, key_pair, salt, class_hash = await deploy_account_details_factory.get()
-
-    account = _account_by_type(
-        account_type=account_type, address=address, client=client, key_pair=key_pair
     )
 
     transaction = await account.sign_deploy_account_transaction(
