+++ starknet_py/tests/e2e/deploy/deployer_test.py
@@ -2,7 +2,7 @@ import pytest
 
 from starknet_py.contract import Contract
 from starknet_py.net.udc_deployer.deployer import Deployer
-from starknet_py.tests.e2e.utils import MAX_FEE
+from starknet_py.tests.e2e.fixtures.constants import MAX_FEE
 from starknet_py.utils.contructor_args_translator import translate_constructor_args
