+++ starknet_py/tests/e2e/fixtures/proxy.py
@@ -5,7 +5,7 @@ import pytest_asyncio
 
 from starknet_py.contract import Contract, DeployResult
 from starknet_py.hash.selector import get_selector_from_name
-from starknet_py.net import AccountClient
+from starknet_py.net.account.account import Account
 from starknet_py.tests.e2e.fixtures.constants import CONTRACTS_PRECOMPILED_DIR, MAX_FEE
 from starknet_py.tests.e2e.fixtures.misc import read_contract
 
@@ -46,27 +46,27 @@ def old_proxy() -> str:
     ]
 )
 async def deploy_proxy_to_contract_oz_argent(
-    request, new_gateway_account_client: AccountClient
+    request, gateway_account: Account
 ) -> DeployResult:
     """
     Declares a contract and deploys a proxy (OZ, Argent) pointing to that contract.
     """
     compiled_proxy_name, compiled_contract_name = request.param
     return await deploy_proxy_to_contract(
-        compiled_proxy_name, compiled_contract_name, new_gateway_account_client
+        compiled_proxy_name, compiled_contract_name, gateway_account
     )
 
 
 @pytest_asyncio.fixture(params=[("oz_proxy_custom_compiled.json", "map_compiled.json")])
 async def deploy_proxy_to_contract_custom(
-    request, new_gateway_account_client: AccountClient
+    request, gateway_account: Account
 ) -> DeployResult:
     """
     Declares a contract and deploys a custom proxy pointing to that contract.
     """
     compiled_proxy_name, compiled_contract_name = request.param
     return await deploy_proxy_to_contract(
-        compiled_proxy_name, compiled_contract_name, new_gateway_account_client
+        compiled_proxy_name, compiled_contract_name, gateway_account
     )
 
 
@@ -74,21 +74,21 @@ async def deploy_proxy_to_contract_custom(
     params=[("oz_proxy_exposed_compiled.json", "map_compiled.json")]
 )
 async def deploy_proxy_to_contract_exposed(
-    request, new_gateway_account_client: AccountClient
+    request, gateway_account: Account
 ) -> DeployResult:
     """
     Declares a contract and deploys a custom proxy pointing to that contract.
     """
     compiled_proxy_name, compiled_contract_name = request.param
     return await deploy_proxy_to_contract(
-        compiled_proxy_name, compiled_contract_name, new_gateway_account_client
+        compiled_proxy_name, compiled_contract_name, gateway_account
     )
 
 
 async def deploy_proxy_to_contract(
     compiled_proxy_name: str,
     compiled_contract_name: str,
-    gateway_account_client: AccountClient,
+    gateway_account: Account,
 ) -> DeployResult:
     """
     Declares a contract and deploys a proxy pointing to that contract.
@@ -96,14 +96,14 @@ async def deploy_proxy_to_contract(
     compiled_proxy = read_contract(compiled_proxy_name)
     compiled_contract = read_contract(compiled_contract_name)
 
-    declare_tx = await gateway_account_client.sign_declare_transaction(
+    declare_tx = await gateway_account.sign_declare_transaction(
         compiled_contract=compiled_contract, max_fee=MAX_FEE
     )
-    declare_result = await gateway_account_client.declare(declare_tx)
-    await gateway_account_client.wait_for_tx(declare_result.transaction_hash)
+    declare_result = await gateway_account.client.declare(declare_tx)
+    await gateway_account.client.wait_for_tx(declare_result.transaction_hash)
 
     declare_proxy_result = await Contract.declare(
-        account=gateway_account_client,
+        account=gateway_account,
         compiled_contract=compiled_proxy,
         max_fee=MAX_FEE,
     )
