+++ starknet_py/tests/e2e/proxy/proxy_test.py
@@ -22,10 +22,10 @@ async def is_map_working_properly(map_contract: Contract, key: int, val: int) ->
 
 
 @pytest.mark.asyncio
-async def test_contract_from_address_no_proxy(account_client, map_contract):
+async def test_contract_from_address_no_proxy(account, map_contract):
     contract = await Contract.from_address(
         address=map_contract.address,
-        client=account_client,
+        provider=account,
     )
 
     assert contract.functions.keys() == {"put", "get"}
@@ -35,15 +35,15 @@ async def test_contract_from_address_no_proxy(account_client, map_contract):
 
 @pytest.mark.asyncio
 async def test_contract_from_address_with_proxy(
-    account_client, deploy_proxy_to_contract_oz_argent
+    account, deploy_proxy_to_contract_oz_argent
 ):
     proxy_contract = await Contract.from_address(
         address=deploy_proxy_to_contract_oz_argent.deployed_contract.address,
-        client=account_client,
+        provider=account,
     )
     proxied_contract = await Contract.from_address(
         address=deploy_proxy_to_contract_oz_argent.deployed_contract.address,
-        client=account_client,
+        provider=account,
         proxy_config=True,
     )
 
@@ -53,31 +53,31 @@ async def test_contract_from_address_with_proxy(
 
 
 @pytest.mark.asyncio
-async def test_contract_from_invalid_address(account_client):
+async def test_contract_from_invalid_address(account):
     with pytest.raises(ContractNotFoundError):
         await Contract.from_address(
             address=123,
-            client=account_client,
+            provider=account,
         )
 
 
 @pytest.mark.asyncio
 async def test_contract_from_address_invalid_proxy_checks(
-    account_client, deploy_proxy_to_contract_custom
+    account, deploy_proxy_to_contract_custom
 ):
     message = "Couldn't resolve proxy using given ProxyChecks"
 
     with pytest.raises(ProxyResolutionError, match=message):
         await Contract.from_address(
             address=deploy_proxy_to_contract_custom.deployed_contract.address,
-            client=account_client,
+            provider=account,
             proxy_config=True,
         )
 
 
 @pytest.mark.asyncio
 async def test_contract_from_address_custom_proxy_check(
-    account_client, deploy_proxy_to_contract_custom
+    account, deploy_proxy_to_contract_custom
 ):
     class CustomProxyCheck(ProxyCheck):
         async def implementation_address(
@@ -96,7 +96,7 @@ async def test_contract_from_address_custom_proxy_check(
 
     contract = await Contract.from_address(
         address=deploy_proxy_to_contract_custom.deployed_contract.address,
-        client=account_client,
+        provider=account,
         proxy_config={"proxy_checks": [CustomProxyCheck()]},
     )
 
@@ -107,11 +107,10 @@ async def test_contract_from_address_custom_proxy_check(
 
 @pytest.mark.asyncio
 async def test_contract_from_address_with_old_address_proxy(
-    new_account_client, old_proxy, map_contract
+    account, old_proxy, map_contract
 ):
-    account_client = new_account_client
     declare_result = await Contract.declare(
-        account=account_client, compiled_contract=old_proxy, max_fee=MAX_FEE
+        account=account, compiled_contract=old_proxy, max_fee=MAX_FEE
     )
     await declare_result.wait_for_acceptance()
     deploy_result = await declare_result.deploy(
@@ -122,11 +121,11 @@ async def test_contract_from_address_with_old_address_proxy(
 
     proxy_contract = await Contract.from_address(
         address=deploy_result.deployed_contract.address,
-        client=account_client,
+        provider=account,
     )
     proxied_contract = await Contract.from_address(
         address=deploy_result.deployed_contract.address,
-        client=account_client,
+        provider=account,
         proxy_config=True,
     )
